Add a force-i option to the dpostgres function
This adds the interactive flag without the TTY. This is mainly for
importing dumps with psql.

Most of the time I use psql directly and for
that I need both the TTY and interactive flags on docker run.

psql receiving a piped dump though needs the interactive flag but not
the TTY one, but psql already gets TTY and interactive by default, so
with force-i I can override that in specific cases.

It's not worth to change the default because most of the time both flags
are what I want when I use psql.