Set histignorespace and globdots opts
histignorespace is to allow for a command beginning with space to not be
recorded in the history (except when it's the last one ran to allow for
easy up arrowing, but it disappears in the next command)

globdots will make * include dot files by default. Without it, we would
need the D glob qualifier.

More info on the glob qualifiers:
http://zsh.sourceforge.net/Doc/Release/Expansion.html#Glob-Qualifiers